refactor: extract all bruker hdf logic
